IRIS-4943 | re-organize structure of page objects for discussions post list and details pages
