+++ src/test/java/com/wikia/webdriver/elements/mercury/pages/discussions/PageWithPosts.java
@@ -1,11 +1,13 @@
 package com.wikia.webdriver.elements.mercury.pages.discussions;
 
 import com.wikia.webdriver.elements.mercury.components.discussions.common.Post;
-import com.wikia.webdriver.elements.mercury.components.discussions.common.PostEntity;
+import com.wikia.webdriver.elements.mercury.components.discussions.common.SignInToFollowModalDialog;
+
 
 public interface PageWithPosts {
 
   Post getPost();
 
-  PostEntity getPostById(String postId);
+  SignInToFollowModalDialog getSignInToFollowModalDialog();
+
 }
